fix/no-overwrite-config-badjson (mycroft-core/pull/2881)
cherry pick of https://github.com/MycroftAI/mycroft-core/pull/2881

Do not overwrite configs with malformed json

If a config file is loaded and is invalid saves to that file will not be
possible until it's manually restored. The store accepts a force flag to
override this protection.

The method returns True if the store succeeded so call sites can
verbally report the issue as well.

Separate tests of LocalConf into new test class