refactor/adapt_ContextManager (mycroft-core/pull/2886)
cherry pick of https://github.com/MycroftAI/mycroft-core/pull/2886

Timed Context Manager

Makes the Mycroft-core Adapt Context Manager depend on the
ContextManager in Adapt. This removes a bunch of duplicated code. The
current state of the class reflects _only_ the modifications done to
integrate the context with the Mycroft intent system.

The key differences from the default Adapt context manager are
- Frames are timedout after a set time
- get_context will only return the latest match
- methods for removing context (all and specific)